requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
boto3==1.28.32
python-dotenv==1.0.0 